string update and message bug

+++ app/src/main/java/com/breadwallet/presenter/fragments/PasswordDialogFragment.java
@@ -476,7 +476,7 @@ public class PasswordDialogFragment extends DialogFragment {
         passcodeEditText.setVisibility(View.GONE);
         info.setVisibility(View.VISIBLE);
         // Get the Resources
-        String message = String.format(getString(R.string.try_again), (int) (waitTime == 0 ? 1 : waitTime),
+        String message = String.format(getString(R.string.try_again), (int) (((int) waitTime) == 0 ? 1 : waitTime),
                 (waitTime == 1 ? getString(R.string.minute) : getString(R.string.minutes)));
 
         info.setText(message);
